+++ lib/core/presentation/pages/event/event_control_panel_page/sub_pages/event_ticket_tier_setting_page/sub_pages/event_create_ticket_tier_page/widgets/ticket_tier_category_form/ticket_tier_category_form.dart
@@ -9,21 +9,19 @@ import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:modal_bottom_sheet/modal_bottom_sheet.dart';
 
-class CreateTicketCategoryBottomsheetForm extends StatefulWidget {
+class TicketTierCategoryForm extends StatefulWidget {
   final Function(String title, String decription) onSubmit;
 
-  const CreateTicketCategoryBottomsheetForm({
+  const TicketTierCategoryForm({
     super.key,
     required this.onSubmit,
   });
 
   @override
-  State<CreateTicketCategoryBottomsheetForm> createState() =>
-      _CreateTicketCategoryBottomsheetFormState();
+  State<TicketTierCategoryForm> createState() => _TicketTierCategoryFormState();
 }
 
-class _CreateTicketCategoryBottomsheetFormState
-    extends State<CreateTicketCategoryBottomsheetForm> {
+class _TicketTierCategoryFormState extends State<TicketTierCategoryForm> {
   String title = '';
   String description = '';
